Refactored friend request api routes
