+++ src/common/page_components/friends/pairedTasks.js
@@ -10,8 +10,8 @@ export function PairedTasks() {
     const [friends, setFriends] = useState()
 
     useEffect(() => {
-        fetch("/api/user/friends/getFriends", {
-            method: "POST"
+        fetch("/api/user/friends", {
+            method: "GET"
         }).then(res => res.json().then(setFriends))
     }, [])
